requirements.txt
paho-mqtt
isort
black
numpy
pyyaml
flake8
mypy
ipython
